participant api

+++ src/main/resources/schema-h2.sql
@@ -1,12 +1,12 @@
 CREATE TABLE public.tri_counts (
-    id int(11) NOT NULL AUTO_INCREMENT,
+    id bigint(11) NOT NULL AUTO_INCREMENT,
     title varchar(50) NOT NULL,
     currency varchar(3) NOT NULL,
     description varchar(200)
 );
 
 CREATE TABLE public.attendees (
-    id int(11) NOT NULL AUTO_INCREMENT,
+    id bigint(11) NOT NULL AUTO_INCREMENT,
     name varchar(50) NOT NULL,
     email varchar(100) NOT NULL,
     into_arrears boolean,
@@ -16,7 +16,7 @@ CREATE TABLE public.attendees (
 );
 
 CREATE TABLE public.expenses (
-    id int(11) NOT NULL AUTO_INCREMENT,
+    id bigint(11) NOT NULL AUTO_INCREMENT,
     title varchar(50) NOT NULL,
     amount double,
     expense_date date,
